requirements.txt
streamlit
tensorflow-cpu
gdown
numpy
pillow
opencv-python-headless